Consolidado

+++ Brito.Sergio.Backend.Acl/Dtos/Fundo.cs
@@ -4,14 +4,14 @@ namespace Brito.Sergio.Backend.Acl.Dtos
 {
     public class Fundo
     {
-        public int CapitalInvestido { get; set; }
-        public int ValorAtual { get; set; }
+        public decimal CapitalInvestido { get; set; }
+        public decimal ValorAtual { get; set; }
         public DateTime DataResgate { get; set; }
         public DateTime DataCompra { get; set; }
-        public int Iof { get; set; }
+        public decimal Iof { get; set; }
         public string Nome { get; set; }
         public decimal TotalTaxas { get; set; }
-        public int Quantity { get; set; }
+        public decimal Quantity { get; set; }
 
         public decimal Ir => (ValorAtual - CapitalInvestido) * (decimal)0.15;
     }
